+++ src/CA/GAIBA.java
@@ -63,7 +63,7 @@ public class GAIBA extends GA{
 		return max;
 	}
 
-	protected double  pixelFitnessFunction(int[][] rules){
+	protected double  pixelFitnessFunction(byte[][] rules){
 		double fitness = 0.0;
 		double maxFitness = 0.0;
 		generator.resetBoard();
@@ -89,7 +89,7 @@ public class GAIBA extends GA{
 
 		return maxFitness;
 	}
-	private double replicationFitnessFunction(int[][] rules, int nrReplicated){
+	private double replicationFitnessFunction(byte[][] rules, int nrReplicated){
 		int partielFitness = 0;
 		int fitness= 0;
 		int maxFitness = 0;
@@ -134,7 +134,7 @@ public class GAIBA extends GA{
 		return maxFitness;
 	}
 	
-	private double evoDevoFitnessFunction(int[][] rules, int nrReplicated){
+	private double evoDevoFitnessFunction(byte[][] rules, int nrReplicated){
 		// development
 		int partielFitness = 0;
 		int fitness= 0;
@@ -300,9 +300,9 @@ public class GAIBA extends GA{
 
 					if(Math.random()<mutateChance){
 						if(j2!=0){
-							nextGen.get(i).getRules()[j][j2] = (int)(Math.random()*5);//FIXED to Von Neumann size 5.
+							nextGen.get(i).getRules()[j][j2] = (byte)(Math.random()*5);//FIXED to Von Neumann size 5.
 						}else{
-							nextGen.get(i).getRules()[j][j2] = (int)(Math.random()*IBARuleSett.ruleName.length);
+							nextGen.get(i).getRules()[j][j2] = (byte)(Math.random()*IBARuleSett.ruleName.length);
 						}
 
 					}
@@ -344,7 +344,7 @@ public class GAIBA extends GA{
 	protected RuleModelIBA crossover(RuleModelIBA dad, RuleModelIBA mom){
 
 		double crossChance = 0.10;
-		int[][] rules =generator.cloneIBARules();//will be overwritten
+		byte[][] rules =generator.cloneIBARules();//will be overwritten
 		boolean dadGenes = true;
 
 		//		System.out.println("dad:" + dad.rulesToBinary());
